pyproject.toml
[tool.poetry]
name = "powerdata-challenge"
version = "0.1.0"
description = ""
authors = ["matheuss0xf <matheussjd9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = {extras = ["standard"], version = "^0.115.8"}
openai = "^1.61.1"
python-dotenv = "^1.0.1"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
rapidfuzz = "^3.12.1"
spacy = "^3.8.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev main.py'
test = 'pytest -s -x --cov=api -vv'
post_test = 'coverage html'